education-agent-2/pyproject.toml
[project]
name = "education-agent-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain[openai]>=0.3.27",
    "langchain-tavily>=0.1.0",
    "langgraph>=0.6.7",
    "langgraph-cli[inmem]>=0.4.2",
    "langsmith>=0.4.28",
    "openai[aiohttp]>=1.107.3",
    "python-dotenv>=1.1.1",
    "streamlit>=1.42.0",
    "tiktoken>=0.11.0",
    "yt-dlp>=2026.3.17",
]

[dependency-groups]
dev = ["ipykernel>=6.30.1", "pytest>=8.0.0"]

[tool.pytest.ini_options]
markers = ["evaluation: AI-as-judge tests requiring OPENAI_API_KEY (deselect with -m 'not evaluation')"]
